requirements.txt
# External packages for Project Management CLI Tool

# Pretty printing for CLI output
rich>=13.0.0

# Date utilities
python-dateutil>=2.8.0

# Testing framework
pytest>=7.0.0
